{{ cookiecutter.project_name }}/requirements/development.txt
-r test.txt
autopep8
dj-database-url==0.5.0
django-debug-toolbar==2.2
django-devserver==0.8.0
django-extensions==3.0.5
django-i18n-helper==0.1.1
django-mail-factory==0.22
django-rename-app==0.1.2
django-template-repl==0.3.0
ipdb==0.13.2
jprq
path.py==11
pep8
pinocchio==0.4.2
pre-commit==2.6.0
pylint
PyVirtualDisplay==0.2.1
PyYAML==5.3.1
recommonmark
sqlparse==0.2.4
transifex-client==0.13.9
virtualenv-tools==1.0
werkzeug==0.16.0
